+++ tests/Response/CountResponseTest.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 namespace Citilink\ExpertSenderApi\Tests\Response;
 
 use Citilink\ExpertSenderApi\Response;
-use Citilink\ExpertSenderApi\Response\DataTablesDeleteRowsPostResponse;
+use Citilink\ExpertSenderApi\Response\CountResponse;
 use PHPUnit\Framework\Assert;
 
 /**
- * DataTablesDeleteRowsPostResponseTest
+ * CountResponseTest
  *
  * @author Nikita Sapogov <sapogov.n@citilink.ru>
  */
-class DataTablesDeleteRowsPostResponseTest extends \PHPUnit_Framework_TestCase
+class CountResponseTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * Test
@@ -21,7 +21,7 @@ class DataTablesDeleteRowsPostResponseTest extends \PHPUnit_Framework_TestCase
     {
         $xml = '<ApiResponse xmlns:xsd="http://www.w3.org/2001/XMLSchema" '
             .'xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"><Count>25</Count></ApiResponse>';
-        $response = new DataTablesDeleteRowsPostResponse(
+        $response = new CountResponse(
             new Response(
                 new \GuzzleHttp\Psr7\Response(
                     200,